requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pytz==2024.1
feedparser==6.0.11
lxml==5.1.0
google-genai==0.3.0
